Convert decimal cols to integer

+++ app/models/profile.rb
@@ -25,14 +25,14 @@ class Profile < ApplicationRecord
   attribute :weekly_report,       :boolean, default: true
   attribute :infinite_tokens,     :boolean, default: false
   attribute :points_claimed,      :integer, default: 0
-  attribute :points_received,     :decimal, default: 0.0
-  attribute :points_sent,         :decimal, default: 0.0
-  attribute :jabs_received,       :decimal, default: 0.0
-  attribute :jabs_sent,           :decimal, default: 0.0
-  attribute :balance,             :decimal, default: 0.0
+  attribute :points_received,     :integer, default: 0
+  attribute :points_sent,         :integer, default: 0
+  attribute :jabs_received,       :integer, default: 0
+  attribute :jabs_sent,           :integer, default: 0
+  attribute :balance,             :integer, default: 0
   attribute :streak_count,        :integer, default: 0
   attribute :tokens_accrued,      :integer, default: 0
-  attribute :tokens_forfeited,    :decimal, default: 0.0
+  attribute :tokens_forfeited,    :integer, default: 0
 
   alias_attribute :points, :points_received
   alias_attribute :jabs, :jabs_received
